backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36

python-dotenv==1.0.1
Werkzeug==3.1.3
gunicorn==23.0.0
flask-socketio==5.4.1
psycopg2-binary==2.9.10
